Implemented: functionality to use remote module in the app and load the components from remote(#863)
