+++ .env.example
@@ -18,3 +18,4 @@ VUE_APP_NOTIF_APP_ID="FULFILLMENT"
 VUE_APP_NOTIF_ENUM_TYPE_ID="NOTIF_FULFILLMENT"
 VUE_APP_FIREBASE_CONFIG={"apiKey": "","authDomain": "","databaseURL": "","projectId": "","storageBucket": "","messagingSenderId": "","appId": ""}
 VUE_APP_FIREBASE_VAPID_KEY=""
+VUE_APP_REMOTE_ENTRY="https://dev-fulfillment-extensions.firebaseapp.com/remoteEntry.js"
