+++ package.json
@@ -4,8 +4,10 @@
   "private": true,
   "description": "An Ionic project",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
+    "devDerve": "vue-cli-service serve --mode=development",
+    "prodServe": "vue-cli-service serve --mode=production",
+    "prodBuild": "vue-cli-service build --mode=production",
+    "devBuild": "vue-cli-service build --mode=development",
     "test:unit": "vue-cli-service test:unit",
     "test:e2e": "vue-cli-service test:e2e",
     "lint": "vue-cli-service lint"
@@ -21,6 +23,7 @@
     "@ionic/core": "^7.6.0",
     "@ionic/vue": "^7.6.0",
     "@ionic/vue-router": "^7.6.0",
+    "@module-federation/enhanced": "^0.7.7",
     "@module-federation/runtime": "^0.7.7",
     "@types/encoding-japanese": "^2.0.5",
     "@types/file-saver": "^2.0.5",
@@ -60,6 +63,6 @@
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.0.3",
     "papaparse": "5.3.1",
-    "typescript": "~4.7.4"
+    "typescript": "~4.9.0"
   }
 }
