+++ src/utils/moduleFederation.ts
@@ -1,0 +1,21 @@
+import { shallowRef } from 'vue';
+import { loadRemote } from '@module-federation/runtime';
+
+const useDynamicImport = ({ module, scope } : any) => {
+  if (!module || !scope) return;
+
+  const loadComponent = async () => {
+    try {
+      const { default: Component } = await loadRemote(`${scope}/${module}`) as any;
+      return shallowRef(Component)
+    } catch (error) {
+      console.error(`Error loading remote module ${scope}/${module}:`, error);
+    }
+  };
+
+  return loadComponent();
+}
+
+export {
+  useDynamicImport
+}
